fix(bridgetracker): address #1866 review findings on exact-settlement search
- Covers now treats the network's empty/initial LER (bridgesynctypes.EmptyLER)
  as covering nothing, instead of 404-ing on GetRootByLER on every retry and
  leaving WaitL1SettledGER stuck forever whenever a bridge's own PreviousLER
  is that empty root (Codex + @claude review).
- EarliestSettlementTxCovering's primary bridge-service path now binary
  searches the settlement history (Covers is monotone over an append-only
  exit tree) instead of walking it linearly one HTTP call per settlement,
  which could outlast the engine's 30s per-tick resolve timeout on a mature
  network. A legacy row with no recorded tx hash is now resolved via a
  single-block FilterLogs at its own BlockNumber before falling back to a
  full genesis-ward scan.
- The L1-log fallback scan is now resumable: it persists a
  *types.SettlementSearchProgress cursor as the step's own Result and
  continues from there on the next tick instead of restarting from
  fromBlock every time it runs out of budget (Codex).
- WaitL1SettledGERResolver.StartDate now pins to EndDate when an earlier
  settlement was swapped in, so the step never reports a negative duration
  against StepCertificatePending's own (later) EndDate (@claude review).
- l1infotreesync.GetVerifiedBatchesPaged computes its offset in uint64 to
  avoid a uint32 overflow on a large/malicious page_number*page_size.

+++ l1infotreesync/processor_verifybatches.go
@@ -140,7 +140,10 @@ func (p *processor) GetVerifiedBatchesPaged(
 		return []*VerifiedBatchWithBlockHash{}, 0, nil
 	}
 
-	offset := (pageNumber - 1) * pageSize
+	// computed in uint64: pageNumber*pageSize can exceed uint32's range for a large/malicious
+	// request, which would otherwise wrap around and silently serve a valid-looking wrong page
+	// instead of an empty one
+	offset := (uint64(pageNumber) - 1) * uint64(pageSize)
 	var verified []*VerifiedBatchWithBlockHash
 	err := meddler.QueryAll(p.db, &verified, `
 		SELECT vb.*, b.hash AS block_hash
